feat: support segmented plot
